% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-missing.R
\name{check_missing}
\alias{check_missing}
\title{Check that the missing data indicator is a scalar}
\usage{
check_missing(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{The provided missing data indicator to check.}

\item{arg}{The name of the argument.}

\item{call}{The call stack.}
}
\value{
The original value (if the checks pass).
}
\description{
The indicator for missing data should be a length 1, non-NULL scalar.
}
\examples{
check_missing(NA)
}
