% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-qmatrix.R
\name{check_qmatrix}
\alias{check_qmatrix}
\title{Title}
\usage{
check_qmatrix(
  x,
  identifier = NULL,
  item_levels = NULL,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{The provided Q-matrix to check.}

\item{identifier}{The provided item identifier, as a character string. If no
item identifier is present, the value should be \code{NULL} (the default).}

\item{item_levels}{Expected}

\item{arg}{The name of the argument.}

\item{call}{The call stack.}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row per item.
}
\description{
Title
}
