% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-class.R
\name{check_S7}
\alias{check_S7}
\title{Check that an object has the expected class}
\usage{
check_S7(x, class, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{The provided object to check.}

\item{class}{The expected class of the object as a string.}

\item{arg}{The name of the argument.}

\item{call}{The call stack.}
}
\value{
Nothing. Called for side effects.
}
\description{
Check that an object has the expected class
}
\examples{
singer <- S7::new_class("singer", properties = list(
  name = S7::class_character,
  birthday = S7::class_Date
))

taylor <- singer(name = "Taylor Swift",
                 birthday = as.Date("1989-12-13"))
check_S7(taylor)
}
