% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdcmchecks-package.R
\docType{package}
\name{rdcmchecks-package}
\alias{rdcmchecks}
\alias{rdcmchecks-package}
\title{rdcmchecks: Common Argument Checks for 'r-dcm' Packages}
\description{
Many packages in the 'r-dcm' family take similar arguments, which are checked for expected structures and values. Rather than duplicating code across several packages, commonly used check functions are included here. This package can then be imported to access the check functions in other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rdcmchecks.r-dcm.org}
  \item \url{https://github.com/r-dcm/rdcmchecks}
  \item Report bugs at \url{https://github.com/r-dcm/rdcmchecks/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Other contributors:
\itemize{
  \item University of Kansas [copyright holder]
  \item Institute of Education Sciences [funder]
  \item Accessible Teaching, Learning, and Assessment Systems [funder]
}

}
\keyword{internal}
